+++ src/adc/reader/speech/__init__.py
@@ -2,4 +2,5 @@ from ._adams import AdamsSpeechReader
 from ._commonvoice import CommonVoiceSpeechReader, COMONVOICE_EXPECTED_HEADER, CommonVoiceDialect
 from ._festvox import FestVoxSpeechReader
 from ._hf_audiofolder import HuggingFaceAudioFolderSpeechReader, HF_AUDIOFOLDER_EXPECTED_HEADER
+from ._piper import PiperSpeechReader
 from ._txt import TxtSpeechReader